test(query-builder): extract shared CodeMirror DOM mocks for jsdom specs
